Refactor: Move is_data_stale to helpers (Fixes #148)
