+++ custom_components/carelink/const.py
@@ -13,7 +13,6 @@ from homeassistant.components.binary_sensor import (
 )
 
 from homeassistant.helpers.entity import EntityCategory
-from homeassistant.util import dt as dt_util
 
 UNAVAILABLE = None
 
@@ -657,25 +656,3 @@ MS_TIMEZONE_TO_IANA_MAP = {
     "Tonga Standard Time": "Pacific/Tongatapu",
     "tzone://Microsoft/Utc": "UTC",
 }
-
-
-def is_data_stale(coordinator_data: dict, sensor_key: str) -> bool:
-    """Check if coordinator data is stale based on last update timestamp.
-    
-    Args:
-        coordinator_data: The coordinator's data dictionary
-        sensor_key: The sensor key (used for logging/debugging)
-    
-    Returns:
-        True if data is stale (older than DATA_STALE_TIMEOUT_HOURS), False otherwise
-    """
-    last_update = coordinator_data.get(SENSOR_KEY_UPDATE_TIMESTAMP)
-    if last_update is None:
-        return True
-    
-    # Calculate time difference
-    now = dt_util.utcnow()
-    time_diff = now - last_update
-    
-    # Return True if data is stale (exceeded threshold)
-    return time_diff >= DATA_STALE_TIMEDELTA
